requirements.txt
streamlit
pandas
psycopg2-binary
Pillow
matplotlib
plotly
